Hoist TokenUsage and the session-vitals family into intendant-core
Extraction-program unlock: TokenUsage (provider.rs) and SessionVitals/
SessionGitVitals/SessionCacheVitals/SessionLimitWindow (types.rs) are
plain shared data with fan-out across the event vocabulary, session
layers, and frontends. They move to intendant-core (usage.rs,
vitals.rs) and re-export at their old mount points, so every existing
crate::provider::TokenUsage / crate::types::Session* path keeps
compiling. This removes the type-level event->provider edge that
blocked the future types+event descent into core.
